requirements.txt
python-dotenv==1.0.0
openai==1.1.0
jsonlines==4.0.0
tqdm==4.66.1
pyyaml==6.0.1
urllib3==1.26.6
huggingface-hub==0.13.4
transformers==4.33.3
datasets==2.11.0
accelerate==0.23.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
jupyter==1.0.0
ipykernel==6.25.1
safetensors==0.3.3
wandb==0.15.12
scikit-learn==1.3.2
matplotlib==3.8.2
torchmetrics==1.2.1
trl==0.7.9
bitsandbytes==0.42.0

# benchmark requirements
stripe==7.10.0
pytelegrambotapi==4.14.1
supabase==2.3.0
slack_sdk==3.26.1
pinecone-client==2.2.4
twilio==8.11.0